Setup database (#4)

+++ Entities/Bucket.cs
@@ -9,7 +9,7 @@ namespace BudgetBucketsAPI.Entities
 		[ForeignKey("Account")]
 		public int AccountId {get; set;}
 		public string Name {get; set;} = String.Empty;
-		public int AmountTotal {get; set;}
+		public decimal AmountTotal {get; set;}
 		public string Note {get; set;} = String.Empty;
 	}
 }
