+++ Entities/CategoryBucket.cs
@@ -11,10 +11,10 @@ namespace BudgetBucketsAPI.Entities
         [ForeignKey("Bucket")]
         public int BucketId {get; set;}
 		public string Name {get; set;} = String.Empty;
-		public int AmountPlanned {get; set;}
-		public int AmountSpent {get; set;}
+		public decimal AmountPlanned {get; set;}
+		public decimal AmountSpent {get; set;}
 		[NotMapped]
-		public int AmountRemaining {get; set;}
+		public decimal AmountRemaining {get; set;}
 	}
 }
